foldes core and redux moved to folder src
