+++ src/components/Dialogs/Dialogs.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
 
-import { ROUTES, I18N} from "../core/constants";
+import { ROUTES, I18N} from "../../core/constants";
 
 import s from "./Dialogs.module.css";
